requirements.txt
google-adk==1.26.0
python-dotenv==1.2.2
fastapi==0.135.1
uvicorn[standard]
sqlalchemy

